ref: Moved `safe_forward` to  `utils.py`
