+++ src/gromo/containers/growing_container.py
@@ -2,31 +2,7 @@ import torch
 
 from gromo.config.loader import load_config
 from gromo.modules.growing_module import GrowingModule, MergeGrowingModule
-from gromo.utils.utils import get_correct_device, global_device
-
-
-def safe_forward(self, input: torch.Tensor) -> torch.Tensor:
-    """Safe Linear forward function for empty input tensors
-    Resolves bug with shape transformation when using cuda
-
-    Parameters
-    ----------
-    input : torch.Tensor
-        input tensor
-
-    Returns
-    -------
-    torch.Tensor
-        F.linear forward function output
-    """
-    assert (
-        input.shape[-1] == self.in_features
-    ), f"Input shape {input.shape} must match the input feature size. Expected: {self.in_features}, Found: {input.shape[1]}"
-    if self.in_features == 0:
-        return torch.zeros(
-            input.shape[0], self.out_features, device=global_device(), requires_grad=True
-        )  # TODO: change to self.device?
-    return torch.nn.functional.linear(input, self.weight, self.bias)
+from gromo.utils.utils import get_correct_device
 
 
 class GrowingContainer(torch.nn.Module):
